move globals
