+++ src/010-pipeline/build.ts
@@ -3,12 +3,12 @@ import { decompileAll } from "@tact-lang/opcode";
 import { writeTypescript } from "../bindings/writeTypescript";
 import { featureEnable } from "../000-config/features";
 import { ConfigProject } from "../000-config/parseConfig";
-import { CompilerContext } from "../context";
+import { CompilerContext } from "./context";
 import { funcCompile } from "../func/funcCompile";
 import { writeReport } from "../generator/writeReport";
 import { getRawAST } from "../grammar/store";
 import files from "../imports/stdlib";
-import { ILogger, Logger } from "../logger";
+import { ILogger, Logger } from "./logger";
 import { PackageFileFormat } from "../packaging/fileFormat";
 import { packageCode } from "../packaging/packageCode";
 import { createABITypeRefFromTypeRef } from "../types/resolveABITypeRef";
