+++ src/browser.ts
@@ -1,6 +1,6 @@
 import { Config, verifyConfig } from "./000-config/parseConfig";
-import { ILogger } from "./logger";
-import { build } from "./pipeline/build";
+import { ILogger } from "./010-pipeline/logger";
+import { build } from "./010-pipeline/build";
 import { createVirtualFileSystem } from "./vfs/createVirtualFileSystem";
 
 export async function run(args: {
