+++ src/generator/writers/writeExpression.spec.ts
@@ -6,7 +6,7 @@ import { WriterContext } from "../Writer";
 import { writeExpression } from "./writeExpression";
 import { openContext } from "../../grammar/store";
 import { resolveStatements } from "../../types/resolveStatements";
-import { CompilerContext } from "../../context";
+import { CompilerContext } from "../../010-pipeline/context";
 import { getParser } from "../../grammar";
 import { getAstFactory } from "../../grammar/ast";
 import { defaultParser } from "../../grammar/grammar";
