+++ src/main.ts
@@ -7,7 +7,7 @@ export {
 
 export { PackageFileFormat } from "./packaging/fileFormat";
 
-export { build } from "./pipeline/build";
+export { build } from "./010-pipeline/build";
 
 export { VirtualFileSystem } from "./vfs/VirtualFileSystem";
 
@@ -15,5 +15,5 @@ export { createVirtualFileSystem } from "./vfs/createVirtualFileSystem";
 
 export * from "./browser";
 export * from "./verify";
-export * from "./logger";
+export * from "./010-pipeline/logger";
 export * from "./error/errors";
