+++ src/verify.ts
@@ -1,10 +1,10 @@
 import normalize from "path-normalize";
 import { Cell } from "@ton/core";
 import { Config, Options } from "./000-config/parseConfig";
-import { ILogger, Logger } from "./logger";
+import { ILogger, Logger } from "./010-pipeline/logger";
 import { PackageFileFormat, run } from "./main";
 import { fileFormat } from "./packaging/fileFormat";
-import { getCompilerVersion } from "./pipeline/version";
+import { getCompilerVersion } from "./010-pipeline/version";
 
 export type VerifyResult =
     | {
